move logic to view model
